\name{double.truncation-package}
\alias{double.truncation-package}
\alias{double.truncation}
\docType{package}
\title{
\packageTitle{double.truncation}
}
\description{
\packageDescription{double.truncation}
}
\details{
Details are seen from the references.
}
\author{
\packageAuthor{double.truncation}

Maintainer: \packageMaintainer{double.truncation}
}
\references{
Hu YH, Emura T (2015) Maximum likelihood estimation for a special exponential family
 under random double-truncation, Computation Stat 30 (4): 1199-229

Emura T, Hu YH, Konno Y (2017) Asymptotic inference for maximum likelihood estimators
 under the special exponential family with double-truncation, Stat Pap 58 (3): 877-909

Dorre A, Emura T (2019) Analysis of Doubly Truncated Data, An Introduction,
 JSS Research Series in Statistics, Springer, Singapore
}
\keyword{ package }