\name{double.truncation-package}
\alias{double.truncation-package}
\alias{double.truncation}
\docType{package}
\title{
Analysis of Doubly-Truncated Data
}
\description{
Likelihood-based inference methods with doubly-truncated data are developed under various models.
 Parametric models from the special exponential family (SEF) are based on
 Hu and Emura (2015) <doi:10.1007/s00180-015-0564-z> and
 Emura, Hu and Konno (2017) <doi:10.1007/s00362-015-0730-y>.
}
\details{
Details are seen from the references.
}
\author{
Takeshi Emura 
Maintainer: Takeshi Emura <takeshiemura@gmail.com>
}
\references{
Hu YH, Emura T (2015) Maximum likelihood estimation for a special exponential family
 under random double-truncation, Computation Stat 30 (4): 1199-229

Emura T, Hu YH, Konno Y (2017) Asymptotic inference for maximum likelihood estimators
 under the special exponential family with double-truncation, Stat Pap 58 (3): 877-909

Dorre A, Emura T (2019) Analysis of Doubly Truncated Data, An Introduction,
 JSS Research Series in Statistics, Springer, Singapore
}
\keyword{ package }